Split file layout accessor
