+++ CMakeLists.txt
@@ -40,6 +40,7 @@ add_library(${PROJECT_NAME}
     src/file.cpp
     src/file_device.cpp
     src/file_layout.cpp
+    src/file_layout_accessor.cpp
     src/free_blocks_allocator.cpp
     src/free_blocks_tree_iterator.cpp
     src/free_blocks_tree.cpp
